requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
flake8==4.0.1
flake8-isort==4.1.1
idna==3.3
isort==5.10.0
mccabe==0.6.1
Pillow==8.4.0
pkg_resources==0.0.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
testfixtures==6.18.3
urllib3==1.26.7
